refactor(Point): 포인트 enum객체 domain으로 이동
