+++ src/main/java/domain/Ladder/Line.java
@@ -1,7 +1,6 @@
 package domain.Ladder;
 
 import domain.LadderGame.SequenceSwapper;
-import domain.util.Point;
 import domain.util.PointGenerator;
 
 import java.util.ArrayList;
