+++ src/main/java/domain/Ladder/Point.java
@@ -1,4 +1,4 @@
-package domain.util;
+package domain.Ladder;
 
 public enum Point {
