+++ src/main/java/domain/util/RandomPointGenerator.java
@@ -1,11 +1,13 @@
 package domain.util;
 
+import domain.Ladder.Point;
+
 import java.util.Random;
 
 public class RandomPointGenerator implements PointGenerator {
-	@Override
-	public Point generate(){
-		boolean point = new Random().nextBoolean();
-		return Point.from(point);
-	}
+    @Override
+    public Point generate() {
+        boolean point = new Random().nextBoolean();
+        return Point.from(point);
+    }
 }
